requirements.txt
llama-index>=0.12.0
llama-index-vector-stores-chroma>=0.4.0
llama-index-embeddings-huggingface>=0.4.0
llama-index-llms-anthropic>=0.6.0
llama-index-readers-file>=0.4.0
chromadb>=0.5.0
pymupdf>=1.24.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
requests>=2.32.0
tqdm>=4.66.0

# Day 3 additions
streamlit>=1.40.0
ragas>=0.2.0
